requirements.txt
requests>=2.31.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
urllib3>=2.0.0
alembic>=1.12.0
pymysql>=1.1.0